Moved broadcast IP settings to settings tab
